fix cert-manager

+++ applications/cert-manager/latest/entrypoint.sh
@@ -9,4 +9,6 @@ HELM_OPTS=${HELM_OPTS:-"
 "}
 
 helm upgrade -i ${NAME} ${CHARTS} -n ${NAMESPACE} --create-namespace ${HELM_OPTS} --wait
-cp -f opt/cmctl /usr/local/bin/
+if [ -f opt/cmctl ]; then
+  cp -f opt/cmctl /usr/local/bin/
+fi
